YourPlaylistDownloader Version 1.0.0
This version of the project is supposed to download only one youtube video at a time. It's version 2.0 will be able to download a whole youtube playlist which is still a work in progress. Help is need from you guys BTW, I have no idea how I will manage; and version 3.0 which will be the last according to my current plan, will be a general download manager which can be used for all your downloads, which I also have no idea how I will manage it.
So, leave your comment and contribute to the project :)